+++ test/correctness/CMakeLists.txt
@@ -341,7 +341,7 @@ tests(
     spirv_ir.cpp
     split_aligned.cpp
     split_aligned_2d.cpp
-    split_aligned_2d_6x6.cpp
+    split_aligned_2d_3x3.cpp
     split_aligned_nested.cpp
     split_aligned_reduction.cpp
     split_by_non_factor.cpp
